test: split gui async status regressions
