+++ crates/localpaste_gui/src/app/tests/mod.rs
@@ -255,5 +255,6 @@ mod keyboard_navigation_audit;
 mod save_and_metadata;
 mod shutdown_behavior;
 mod state_basics;
+mod version_async_status;
 mod version_modal_caching;
 mod virtual_editor_behaviors;
